feat: updateBookingRevenueStats

+++ src/main/java/com/app/CourtReservationSystem/dto/statistic/SystemStatisticResponse.java
@@ -8,17 +8,17 @@ import java.util.List;
 
 @Setter @Getter
 public class SystemStatisticResponse {
-    private Integer totalUsers;
-    private Integer totalOrganisations;
-    private Integer totalOrganisationsOpening;
-    private Integer totalOrganisationsMaintaining;
-    private Integer totalOrganisationsClosed;
+    private Long totalUsers;
+    private Long totalOrganisations;
+    private Long totalOrganisationsOpening;
+    private Long totalOrganisationsMaintaining;
+    private Long totalOrganisationsClosed;
     private Double totalRevenueBooking; //Tong doanh thu cua tat ca booking (Theo thời gian)
-    private Integer bookingTimes; //So tong booking (Theo thời gian)
+    private Long bookingTimes; //So tong booking (Theo thời gian)
     private List<RevenueBookingLocation> revenueBookingLocations;
     private Double totalRevenueOrder; //Tong doanh thu don hang
-    private Integer successOrderCount;
-    private Integer processingOrderCount;
-    private Integer failedOrderCount;
+    private Long successOrderCount;
+    private Long processingOrderCount;
+    private Long failedOrderCount;
     private List<ProductResponse> topProduct;
 }
